+++ pkg/sites/defaults/defaults.go
@@ -1,36 +1,41 @@
 package defaults
 
 import (
-	hackernews "github.com/go-go-golems/scraper/pkg/sites/hackernews"
+	sitemanifest "github.com/go-go-golems/scraper/pkg/sites/manifest"
+	"github.com/go-go-golems/scraper/pkg/sites/registry"
+
+	"github.com/go-go-golems/scraper/pkg/sites/hackernews"
 	"github.com/go-go-golems/scraper/pkg/sites/jsdemo"
 	"github.com/go-go-golems/scraper/pkg/sites/nereval"
-	"github.com/go-go-golems/scraper/pkg/sites/registry"
 	"github.com/go-go-golems/scraper/pkg/sites/slashdot"
 )
 
+// NewRegistry creates a registry with all built-in sites registered.
 func NewRegistry() (*registry.Registry, error) {
 	ret := registry.New()
-	if err := Register(ret); err != nil {
+
+	if err := ret.Register(jsdemo.Definition()); err != nil {
 		return nil, err
 	}
+	if err := ret.Register(hackernews.Definition()); err != nil {
+		return nil, err
+	}
+	if err := ret.Register(slashdot.Definition()); err != nil {
+		return nil, err
+	}
+	if err := ret.Register(nereval.Definition()); err != nil {
+		return nil, err
+	}
+
 	return ret, nil
 }
 
-func Register(r *registry.Registry) error {
-	if r == nil {
-		r = registry.New()
+// LoadExternalSites loads additional sites from a directory of site manifests.
+// Each subdirectory containing a site.yaml is registered as a site.
+// Returns nil if dir is empty.
+func LoadExternalSites(r *registry.Registry, dir string) error {
+	if dir == "" {
+		return nil
 	}
-	if err := hackernews.Register(r); err != nil {
-		return err
-	}
-	if err := slashdot.Register(r); err != nil {
-		return err
-	}
-	if err := jsdemo.Register(r); err != nil {
-		return err
-	}
-	if err := nereval.Register(r); err != nil {
-		return err
-	}
-	return nil
+	return sitemanifest.RegisterDir(r, dir)
 }